Added OpenGL vertex formats.

+++ Render/src/Resources/ResourceManager.c
@@ -140,7 +140,7 @@ void dsResourceManager_reportStatistics(const dsResourceManager* resourceManager
 		(double)resourceManager->textureMemorySize/(1024.0*1024.0));
 	DS_PROFILE_STAT("ResourceManager", "Renderbuffers", resourceManager->renderbufferCount);
 	DS_PROFILE_STAT("ResourceManager", "Renderbuffer memory (MB)",
-		resourceManager->renderbufferMemorySize);
+		(double)resourceManager->renderbufferMemorySize);
 	DS_PROFILE_STAT("ResourceManager", "Framebuffers", resourceManager->framebufferCount);
 	DS_PROFILE_STAT("ResourceManager", "Fences", resourceManager->fenceCount);
 	DS_PROFILE_STAT("ResourceManager", "Shader modules", resourceManager->shaderModuleCount);
